(feat) moved userHelpers into ModelHelpers
